cleanup, dynamic config, and working tests except for reset
