+++ chasm/lib/workflow/config.go
@@ -11,6 +11,8 @@ type Config struct {
 	maxLinksPerRequest                dynamicconfig.IntPropertyFnWithNamespaceFilter
 	linkMaxSize                       dynamicconfig.IntPropertyFnWithNamespaceFilter
 	enableSignalWithStartFromWorkflow dynamicconfig.BoolPropertyFnWithNamespaceFilter
+	enableGetWorkflowExecutionResult  dynamicconfig.BoolPropertyFnWithNamespaceFilter
+	maxCallbacksPerWorkflow           dynamicconfig.IntPropertyFnWithNamespaceFilter
 }
 
 func NewConfig(dc *dynamicconfig.Collection) Config {
@@ -20,5 +22,7 @@ func NewConfig(dc *dynamicconfig.Collection) Config {
 		maxLinksPerRequest:                dynamicconfig.FrontendMaxLinksPerRequest.Get(dc),
 		linkMaxSize:                       dynamicconfig.FrontendLinkMaxSize.Get(dc),
 		enableSignalWithStartFromWorkflow: dynamicconfig.EnableSignalWithStartFromWorkflow.Get(dc),
+		enableGetWorkflowExecutionResult:  dynamicconfig.EnableGetWorkflowExecutionResult.Get(dc),
+		maxCallbacksPerWorkflow:           dynamicconfig.MaxCallbacksPerWorkflow.Get(dc),
 	}
 }
