+++ common/dynamicconfig/constants.go
@@ -2995,6 +2995,12 @@ instead of the previous HSM backed implementation.`,
 		`Controls whether signal with start from workflow is enabled.`,
 	)
 
+	EnableGetWorkflowExecutionResult = NewNamespaceBoolSetting(
+		"history.enableGetWorkflowExecutionResult",
+		false,
+		`Controls whether GetWorkflowExecutionResult is enabled.`,
+	)
+
 	EnableCHASMSignalBacklinks = NewNamespaceBoolSetting(
 		"history.enableCHASMSignalBacklinks",
 		false,
