+++ service/frontend/workflow_handler.go
@@ -27,7 +27,6 @@ import (
 	updatepb "go.temporal.io/api/update/v1"
 	workerpb "go.temporal.io/api/worker/v1"
 	workflowpb "go.temporal.io/api/workflow/v1"
-	"go.temporal.io/api/workflownexusservice/v1"
 	"go.temporal.io/api/workflowservice/v1"
 	batchspb "go.temporal.io/server/api/batch/v1"
 	deploymentspb "go.temporal.io/server/api/deployment/v1"
@@ -7342,8 +7341,4 @@ func (wh *WorkflowHandler) TerminateNexusOperationExecution(_ context.Context, _
 
 func (wh *WorkflowHandler) DeleteNexusOperationExecution(_ context.Context, _ *workflowservice.DeleteNexusOperationExecutionRequest) (_ *workflowservice.DeleteNexusOperationExecutionResponse, retError error) {
 	return nil, serviceerror.NewUnimplemented("DeleteNexusOperationExecution is not yet implemented")
-}
-
-func (wh *WorkflowHandler) GetWorkflowExecutionResult(_ context.Context, _ *workflownexusservice.GetWorkflowExecutionResultRequest) (_ *workflownexusservice.GetWorkflowExecutionResultResponse, retError error) {
-	return nil, serviceerror.NewUnimplemented("GetWorkflowExecutionResult gRPC endpoint is not yet implemented")
 }
